Move Resolver 'camera' react code to a hook
Move camemra logic to `useCamera`
