+++ x-pack/plugins/endpoint/public/embeddables/resolver/view/use_autoupdating_client_rect.ts
