+++ x-pack/plugins/endpoint/public/embeddables/resolver/view/use_nonpassive_wheel_handler.ts
